+++ webscout/Provider/Deepinfra.py
@@ -16,6 +16,8 @@ class DeepInfra(Provider):
     required_auth = False
     AVAILABLE_MODELS = [
         "moonshotai/Kimi-K2-Instruct",
+        "moonshotai/Kimi-K2-Thinking",
+        "MiniMaxAI/MiniMax-M2",
         "Qwen/Qwen3-Next-80B-A3B-Instruct",
         "Qwen/Qwen3-Next-80B-A3B-Thinking",
         "moonshotai/Kimi-K2-Instruct-0905",